image_gen/requirements.txt
crewai
google-generativeai
langchain-google-genai
python-dotenv
requests
pillow